move interface to a more-appropriate file
